requirements.txt
aiogram==3.10.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
redis==5.0.7
apscheduler==3.10.4
fluent.runtime==0.4.0
python-dotenv==1.0.0
requests==2.31.0
gspread>=6.0.0
google-auth>=2.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
